Step 1 : strongly simplify addon design & loading (preserve tests)
