--- addons/attach/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "xterm.attach",
-  "main": "attach.js",
-  "private": true
-}
