--- addons/fit/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "xterm.fit",
-  "main": "fit.js",
-  "private": true
-}
