--- addons/fullscreen/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "xterm.fullscreen",
-  "main": "fullscreen.js",
-  "private": true
-}
