--- addons/linkify/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "xterm.linkify",
-  "main": "linkify.js",
-  "private": true
-}
