+++ src/xterm.js
@@ -664,24 +664,6 @@ var Viewport          = require('./lib/viewport');
     };
 
 
-    /**
-     * Attempts to load an add-on using CommonJS or RequireJS (whichever is available).
-     * @param {string} addon The name of the addon to load
-     * @static
-     */
-    Terminal.loadAddon = function(addon, callback) {
-      if (typeof exports === 'object' && typeof module === 'object') {
-        // CommonJS
-        return require(__dirname + '/../addons/' + addon);
-      } else if (typeof define == 'function') {
-        // RequireJS
-        return require(['../addons/' + addon + '/' + addon], callback);
-      } else {
-        console.error('Cannot load a module without a CommonJS or RequireJS environment.');
-        return false;
-      }
-    };
-
 
     /**
      * XTerm mouse events
